requirements.txt
Flask
google-cloud-logging
google-cloud-storage
gunicorn
pandas
python-dotenv